+++ Point.cpp
@@ -15,7 +15,7 @@ Point::Point(int NumRows, int NumCols) : Matrix(4, 1)
         throw std::invalid_argument("size of a Point must be 4x1");
 }
 
-Point::Point(float X, float Y, float Z) : Matrix(4, 1)
+Point::Point(double X, double Y, double Z) : Matrix(4, 1)
 {
     this->Set(0, 0, X);
     this->Set(1, 0, Y);
@@ -44,40 +44,40 @@ std::ostream &operator<<(std::ostream &os, const Point &P)
 
 TEST_CASE("testing the Point")
 {
-    Point v = Point(4.3f, -4.2f, 3.1f);
-    CHECK(v == Matrix(4.3f, -4.2f, 3.1f, 1.f));
+    Point v = Point(4.3, -4.2, 3.1f);
+    CHECK(v == Matrix(4.3, -4.2, 3.1, 1.f));
 }
 
 TEST_CASE("multiplying a point with a scalar")
 {
-    Point v = Point(1.f, -1.f, 3.f);
+    Point v = Point(1., -1., 3.f);
     Matrix M = Matrix(4, 1);
-    M(0, 0) = 2.f;
-    M(1, 0) = -2.f;
-    M(2, 0) = 6.f;
-    M(3, 0) = 1.f;
+    M(0, 0) = 2.;
+    M(1, 0) = -2.;
+    M(2, 0) = 6.;
+    M(3, 0) = 1.;
 
     CHECK(v * 2 == M);
 }
 
 TEST_CASE("subtracting two points")
 {
-    Point A = Point(3.f, 2.f, 1.f);
-    Point B = Point(5.f, 6.f, 7.f);
-    CHECK((A - B) == Vector(-2.f, -4.f, -6.f));
+    Point A = Point(3., 2., 1.f);
+    Point B = Point(5., 6., 7.f);
+    CHECK((A - B) == Vector(-2., -4., -6.f));
 }
 
 TEST_CASE("add one point to one vector")
 {
-    Point A = Point(3.f, 2.f, 1.f);
-    Vector B = Vector(5.f, 6.f, 7.f);
-    CHECK((A + B) == Point(8.f, 8.f, 8.f));
-    CHECK((B + A) == Point(8.f, 8.f, 8.f));
+    Point A = Point(3., 2., 1.f);
+    Vector B = Vector(5., 6., 7.f);
+    CHECK((A + B) == Point(8., 8., 8.f));
+    CHECK((B + A) == Point(8., 8., 8.f));
 }
 
 TEST_CASE("subtracting a vector from a point")
 {
-    Point A = Point(3.f, 2.f, 1.f);
-    Vector B = Vector(5.f, 6.f, 7.f);
-    CHECK((A - B) == Point(-2.f, -4.f, -6.f));
+    Point A = Point(3., 2., 1.f);
+    Vector B = Vector(5., 6., 7.f);
+    CHECK((A - B) == Point(-2., -4., -6.f));
 }
